requirements.txt
pdfplumber
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slack-sdk
python-dotenv
